chore: new schema package
